Rename SdlConnection to session in javaSE
